User supervisord to handle the two processes
The docker container consists for two processes: xl2tpd and ipsec. Previously,
the former was launched in the background and the latter was kept in the
foreground.

This has two issues:
   1. You can't gracefully stop the docker container anymore because none of the
      two processes receive the terminate signal from docker.

   2. If xl2tpd fails in the background, this goes unnoticed and the docker-container
      does not stop.

This commit replaces this behaviour: Supervisord is used to control the two
processes. If one of them fail, the whole docker container will fail. Also the
stdout and stderr is collected and aggregated and can therefor be viewed in the
docker logs.